chore: make `export_builder_and_near_abi_versions` and `get_coerced_builder_version` methods on `VersionMismatch`
