+++ cargo-near-build/src/types/near.rs
@@ -7,12 +7,13 @@ use crate::{ArtifactType, WASM};
 
 // TODO: make non-pub
 pub mod abi;
+pub mod build;
 
 pub struct CompilationArtifact<T: ArtifactType = WASM> {
     pub path: Utf8PathBuf,
     pub fresh: bool,
     pub from_docker: bool,
-    pub cargo_near_version_mismatch: VersionMismatch,
+    pub cargo_near_version_mismatch: build::VersionMismatch,
     pub artifact_type: PhantomData<T>,
 }
 
@@ -39,34 +40,5 @@ impl SHA256Checksum {
 
     pub fn to_base58_string(&self) -> String {
         bs58::encode(&self.hash).into_string()
-    }
-}
-
-#[derive(Debug)]
-pub enum VersionMismatch {
-    Some {
-        environment: String,
-        current_process: String,
-    },
-    None,
-    UnknownFromDocker,
-}
-
-impl std::fmt::Display for VersionMismatch {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Some {
-                environment,
-                current_process,
-            } => {
-                write!(
-                    f,
-                    "`cargo-near` version {} -> `cargo-near` environment version {}",
-                    current_process, environment
-                )
-            }
-            Self::None => write!(f, "no `cargo-near` version mismatch in nested builds detected",),
-            Self::UnknownFromDocker => write!(f, "it's unknown if `cargo-near` version mismatch has occurred in docker build environment",),
-        }
     }
 }
